Update client policies, package versions, and tests
Former-commit-id: 1590bdcf4c197c79f7e2f317f8cebca09d0c5af5

+++ VpnHood.AccessServer.Persistence/Models/ServerStatusModel.cs
@@ -9,7 +9,7 @@ public class ServerStatusBaseModel
     public required int TcpConnectionCount { get; set; }
     public required int UdpConnectionCount { get; set; }
     public required long? AvailableMemory { get; set; }
-    public required long? AvailableSwapMemoryMb { get; set; }
+    public required int? AvailableSwapMemoryMb { get; set; }
     public required byte? CpuUsage { get; set; }
     public required int ThreadCount { get; set; }
     public required long TunnelSendSpeed { get; set; }
